Add new public method: hideKeyboard
I have found that when I was using DAKeyboardControl
I have wanted to programmatically hide the keyboard 
created by DAKeyboardControl.

This commit adds a public method to the category,
allowing you to send a message to the receiving UIView
to hide the keyboard.

I have also replaced the occurrences in the code 
that were hiding the keyboard to use the new public method.