Fixed animation options
First and foremost, the animations were not using the `UIViewAnimationOptionBeginFromCurrentState` option, which caused it to jump, for instance when doing the following, which forces autocorrect to complete:

    [input resignFirstResponder];
    [input becomeFirstResponder];

Second, the animation curve from the keyboard info was being used as an option, which are different types. This was mostly benign, since it's always ease in/out, which in the curve type is 0, which is nothing in the options type, but for safety a conversion method was added.